// front-end/lib/types.ts
export type DietType =
  | "breakfast"
  | "lunch"
  | "dinner"
  | "snack"
  | "afternoon-snack"
  | "pre-workout"
  | "post-workout";
// ...
export interface MealFoodItem {
  id: string;
  foodId: string;
  foodName: string;
  servings: number;
  calories: number;
  protein: number;
  carbs: number;
  fats: number;
  servingSize: string;
}

// Refeição individual
export interface Meal {
  id: string;
  name: string;
  calories: number;
  protein: number;
  carbs: number;
  fats: number;
  completed?: boolean;
  type: DietType;
  time?: string;
  image?: string;
  ingredients?: string[];
  foods?: MealFoodItem[]; // Alimentos adicionados dinamicamente
}